Add admin functionality: enhance UI components, update API routes, and implement student management features
