+++ ContentWebApp/src/components/allContent.css
@@ -115,6 +115,8 @@
   min-width: 180px;
   z-index: 1000;
   overflow: hidden;
+  padding: 15px 0;
+  font-size: 20px;
 }
 
 .dropdown-item {
@@ -386,13 +388,14 @@
 
 .registration-title {
   margin: 0;
-  color: #0f172a;
+  color: #ffffff;
 }
 
 .input-field {
   width: 100%;
   border-radius: 10px;
   border: 1px solid #cbd5f5;
+  background-color: #ffffff;
   padding: 12px;
   font-size: 15px;
 }
