+++ backend-server/src/index.js
@@ -10,6 +10,7 @@ const {port} = require("./config/env");
 const authenticateToken = require('./auth/authenticateToken');
 const callRouter = require("./routes/callRouter.js");
 const teacherRouter = require("./routes/teacherRouter.js");
+const v1TeacherRouter = require("./routes/v1TeacherRouter.js");
 const contentRouter = require("./routes/contentRouter");
 const classRoomRouter = require("./routes/classRouter.js");
 const userRouter = require("./routes/userRouter.js");
@@ -46,6 +47,7 @@ app.use("/log", authenticateToken, logRouter);
 app.use("/user", authenticateToken, userRouter);
 
 app.use("/teacher", teacherRouter);
+app.use("/v1/teacher", v1TeacherRouter);
 app.use("/tenant", tenantRouter);
 if (require.main === module) {
   mongo()
